+++ src/events/SDL_mouse_c.h
@@ -58,7 +58,7 @@ typedef struct
     int xdelta;
     int ydelta;
     int last_x, last_y;         /* the last reported x and y coordinates */
-    Uint8 buttonstate;
+    Uint32 buttonstate;
     SDL_bool relative_mode;
     /* the x and y coordinates when relative mode was activated */
     int original_x, original_y;
